feat: hook for theme change
