+++ apps/web/src/hooks/index.ts
@@ -23,4 +23,4 @@ export * from './useInlineComponent';
 export * from './useHoverOverItem';
 export * from './useHover';
 export * from './useBootIntercom';
-export * from './useTrackThemeChange';
+export * from './useThemeChange';
